requirements.txt
appdirs==1.4.3
astroid==2.4.1
attrs==19.3.0
autopep8==1.5.2
black==19.10b0
click==7.1.1
colorama==0.4.3
entrypoints==0.3
flake8==3.7.9
isort==4.3.21
keyring==21.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
pathspec==0.7.0
pycodestyle==2.5.0
pyflakes==2.2.0
pyftpsync==3.1.0
pylint==2.5.2
PyYAML==5.3.1
regex==2020.4.4
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
wrapt==1.12.1
